Create Lab1_Exercise2_DAC_&_Comparator
The potentiometer on the Curiosity board rotates between approximately 7 o'clock (all the way anti-clockwise, Vout(RC0) = 0 V) to 5 o'clock (all the way clockwise, Vout(RC0) = 3.3 V) positions. Program the microcontroller to change LED color based on the rotary position of the potentiometer. When the potentiometer is between 7 o'clock and 12 o'clock, the LED should be green. When it is between 12 o'clock and 5 o'clock, the LED should be red. You can assume that the 12 o'clock positions correspond to Vout on the RC0 pin being ~1.65 V, respectively. Use the DAC to generate
the appropriate references voltage for Comparator 1 to compare against the voltage on pin RC0. You will need to use a wire to connect pin RC0 on the J3 header to an appropriate comparator input pin.