Create Lab2_Exercise4_LED_intensity_control_with_timer
This exercise needs to be done in a group of 4 using 4 microcontrollers. Set the oscillator to 8 MHz (Fosc) and the timer clock to the instruction clock (Fosc/2), prescaler of 1 for each board. Using the PR1 register and the value of the TMR1 register,
Group member 1 program blue LED to be on for 8 ms, off for 2 ms, then repeat.
Group member 2 program blue LED to be on for 6 ms, off for 4 ms, then repeat.
Group member 3 program blue LED to be on for 4 ms, off for 6 ms, then repeat.
Group member 4 program blue LED to be on for 2 ms, off for 8 ms, then repeat.
Compare the intensities of the LEDs on the 4 boards. This technique of using different duty cycles to control LED intensity is called pulse width modulation. It is also used for other applications such as motor speed control.