Create Lab3_Exercise1_Reception_of_data_sent_from_computer_to_microcontroller
Run the MATLAB function as serialWrite('COMX'); where COMX is the name of the port in the device manager. This script will ask you to input a character and transmit it over the UART as the
corresponding ASCII code. On the microcontroller, wait for a character to be received (URXDA) and read it from U1RXREG. If the character ‘r’ is received, turn on just the red LED. If the character ‘g’ is received, turn on just the green LED. If the character ‘b’ is received, turn on just the blue LED. If the
character ‘n’ is received, turn off all LEDs. All other characters should not have any effect. After doing this, go back to waiting for another character from the computer.