Create Lab3_Exercise2_Transmission_of_data_from_microcontroller_to_computer
Run the MATLAB function as serialRead('COMX'); where COMX is the name of the port in the device manager. This script will read data sent to the computer and display it. Wait till either S1 or
S2 is pressed. If S1 was pressed, send the character ‘1’ via UART (by writing the ASCII code for ‘1’ to U1TXREG). If S2 was pressed send the character ‘2’. Once a character is sent, wait for the
transmission to complete (TRMT) and then go back to waiting for a switch press.